Progress toward handling node sets defined in an Exodus input mesh.
Node sets are now preserved when converting a hex
mesh from an Exodus input file into a sphere mesh
for use in Peridigm.  A Peridigm node set is created
for each node set defined in the input Exodus mesh
file.  When creating the sphere elements, if a node
in a given hex element is part of an Exodus node set,
then the node at the center of the newly-created sphere
element will be part of the corresponding Peridigm
node set.

One thing to be aware of is that node sets read from
parallel Exodus files will include only locally-owed
nodes.  It would be possible to implement some parallel
communication such that all processes would be aware
of all the nodes in all node sets, regardless of whether
the node is locally owned, but this logic is not currently
in place.

Created a function Peridigm::initializeNodeSets()
that combines node set definitions from the input
deck and from an Exodus input mesh.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@838 d9297710-12f9-482b-b11f-fb677dc88bbb