Refactored Peridigm_NS::Material and Peridigm_NS::LinearElasticIsotropicMaterial for switch to new libraries.
The unit test utPeridigm_LinearElasticIsotropicMaterial
builds, runs, and passes.

IMPORTANT:
In the old version of Field.h, there was a scalar
point spec called DAMAGE, which was the percent of
broken bonds, and a scalar bond spec called BOND_DAMAGE,
which was the damage value for a specific bond.

In the new version of Field.h, the Field_ENUM::Type
enumeration includes both DAMAGE and BOND_DAMAGE, which
is consistent with the old Field.h.  However, the new
version of Field.h includes a definition for only the
scalar bond spec DAMAGE.  This conflicits with the old
Field.h, where DAMAGE was a scalar point spec, and
BOND_DAMAGE was a scalar bond spec.

I updated the new Field.h to match what was in the old
Field.h.  Now there is a scalar point spec called DAMAGE
and a scalar bond spec called BOND_DAMAGE.  Both are
required for the Peridigm material models.  There is some
danger that switching the definition of DAMAGE in the
new Field.h from a bond spec to a point spec could break
something outside Peridigm that uses this file.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@854 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 2db4cf4c5752288c7e93064ca76158f3e206dc61