Fixed utPeridigm_State unit test.
The problem was a result of confusion between the
scalar point spec DAMAGE and the scalar bond spec
BOND_DAMAGE.

Also, added check to State class that throws an
exception if you try to register an inconsistent
state (e.g. if you try to put a scalar bond spec
into the scalar point multivector).



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@857 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 5eac35bd27a8ddacf6065a9a3f2630564c3201a8