Added a print statement to the QuickGridData.
Also added a new unit test that is exactly a case 
that is in utPeridigm.

I'm still a little confused as to what is going on 
but I am slowly narrowing this down to a point where 
I will be able to say that there is nothing wrong with
the GridData:)

I modified the test slightly because it was set in a 
pathological way with the horizon exactly on the cell 
spacing. 




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@869 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: da253777e72a7c631f546ea2eea073509244d158