Adds bond volume calculation.
Notes:
To properly use the bond volume, we have to 
develop a new pattern.  Briefly, the issue is 
as follows:

In the internal force calculation for point P, 
we loop over neighboring points Q.  From P to Q 
we have the bond volume.  However, upon inspection, 
Mike and I realized that we also need the bond volume 
from Q to P -- at least that is how we currently have 
things set up.

An incomplete solution amounts to creating another 
bond variable that has exactly the same topology as
all bond variables except that it stores the Q to P 
bond volume.  On processor, how this is collected is 
a problem that needs to be solved.




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@876 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 86cb3b2d32643d4528247ba95df798f4d3f8888b