Fixed sneaky bug in PDNEIGH::NeighborhoodList that caused failures on Intel build (Redsky).
The issue was in the creation of the maps, for example:

epetra_block_maps[Epetra_MapTag(OWNED,1)] =
  create_Epetra_BlockMap(Epetra_MapTag(OWNED,1));

The problem is that the function create_Epetra_BlockMap() actually
checks epetra_blocks_maps to see if there is an entry for
Epetra_MapTag(OWNED,1).  If there is an entry, it returns it.
If not, it creates a new map.

On Intel builds, apparently the above command results in the
creation (initialization?) of a map entry for Epetra_MapTag(OWNED,1)
before the call to create_Epetra_BlockMap() executes.  Thus,
when the function create_Epetra_BlockMap() checks for the existence
of the entry, it finds one and returns it.  But this entry is a smart
pointer to nothing, and the code dies downstream.

The fix is just to create the map on a separate command line and then
add it to the map:

shared_ptr<Epetra_BlockMap> owned1DMap =
  create_Epetra_BlockMap(Epetra_MapTag(OWNED,1));

epetra_block_maps[Epetra_MapTag(OWNED,1)] = owned1DMap;



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@879 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: a5c15be402d0f1de2b20ec67fa5393e30b2313b0