Added check for NaN and Inf in global force vector and global contact force vector.
The idea here is that if we get a NaN in the
force evaluation, we'd like to know right away.
I was trying a contact problem with (I believe)
bad input values that result in huge huge
forces.  The code responded by crashing in
a difficult-to-find way downstream.  The
checks added in this commit will catch problems
much sooner.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@884 d9297710-12f9-482b-b11f-fb677dc88bbb