The first of what will be many commits toward handling multiple material blocks.
The getMap() and getOverlapMap() functions in
the discretization classes have been renamed
getGlobalMap() and getGlobalOverlapMap().  When
handling multiple material blocks, there will
be a global set of maps (all points, regardless
of material, applicable to mothership vectors),
and there will be a set of maps for each
material (i.e., there will be a DataManager
for each material).

Also fixed an error in parallel PdQuickGrid
unit test.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@888 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 4d5ade835e323ab1dc0e15b977de3f90f30c91b7