Incremental check-in toward handling multiple material blocks.
The main change in this check-in is replacing
the DataManager in Peridigm with a vector of
DataManagers.

The game plan is to create a Block container
class to hold everything that is associated with
a given block.  Namely, the MaterialModel,
ContactModel, DataManager, and Importers for
the DataManagers.  Currently each of these
is stored directly in the main Peridigm class
and it's unruly.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@889 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: e145329cc8f43836fa6e9c63bcabdaa326ab07aa