First steps toward actually hooking up new code for multiple material blocks.
Currently a number of calculations are being done
twice, once on the old data structures and once
on the new data structures (allows for comparisons).
There are #defines (currently turned off) that
switch over to actually use the data in the new
structures.

Once everything is working with the new structures,
I'll rip out the old ones and the run time will
come back down by a factor of two.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@898 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 3aced91b956c98ac605ad94d906bbea8969bf030