New unit test for naive quadrature convergence study.  This test dumps an asci file with the following columns:
n h m_n ed^2_n

where 
n is the number of cells on an edge,
h is the cell size
m_n is the weighted volume
ed^2_n deviatoric energy associated with pure shear

The unit test diffs this asci file with a gold copy.

Updated the existing bond volume convergence test to do 
the same as above.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@907 d9297710-12f9-482b-b11f-fb677dc88bbb