Progress toward multiple material blocks.
My current approach is to use a #define
MULTIPLE_BLOCKS to basically switch between
two versions of the code:  The old version
where Peridigm holds a single DataManager,
Material, etc., and the new version where
everything is kept inside Block objects.

The checked-in version of the code is
using the old, single-material approach.

Two areas that need significant work prior
to turning on multiple material blocks are
output and rebalance.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@920 d9297710-12f9-482b-b11f-fb677dc88bbb