Progress toward multiple material blocks.
Implemented functionality in Block class to
create block-specific maps and neighborhood
lists based on the global maps, global neighborhood
lists, and a vector that contains the block ID
of each element.  This takes advantage of the fact
that the block-specific data are a subset of the
global data (with the added complication that all
the local IDs are different).

This design is different from the initial design
for multiple blocks wherein the Discretization
objects were responsible for creating block-specific
maps and neighborhood lists.  The big selling point
of the new approach is that it can be applied
following a rebalance.  The rebalance operation
creates new global maps and neighborhood data,
and this is then fed to the Blocks so they can
update their maps and neighborhood data, etc.

Also, this approach keeps the Discretization
objects cleaner.  The only block-related thing
they have to do is return a vector that has the
block ID of each element.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@924 d9297710-12f9-482b-b11f-fb677dc88bbb