This check-in activates the new multiple-block code by setting #define MULTIPLE_BLOCKS.  The old code is still there, next up is to remove it.
The disk_impact example problem was reworked
because the original version consisted of
two blocks.  This worked before because block
information was ignored.  The new code
recognizes the blocks, but is not fully
capable of handling them.  So, the Cubit
journal file has been altered to produce
only one block.

The disk_impact example has been checked
for consistency among the old and new code,
on one and two processors.  All cases match
as expected (exactly when the number of
processors is the same, and within tolerance
when the number of processors is different).

A Block bug was also fixed in this check-in,
the Importers were not being properly reset
following rebalance.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@933 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: ec4dea342f13701a504f0440e7418136f1b135c2