Reverting to r960. This will fix build issues against Trilinos 10.8.
The new "TEST_FOR_EXCEPT" macros break backwards compatibility with Trilinos 10.8, so we can't use them until the next Trilinos release.
The old macros (as of this afternoon) now work again in the Trilinos VOTD, but issue a depreciation warning 
at compile time, resulting in hundreds of warnings when building Peridigm. 




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@962 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 0a1baced3d08bd328492b21efb12d64cfdfbdc82