First attempt to resolve compiler errors encountered with Intel 10.1 on ranger.
The issue is related to the header memory.hpp, which
contains smart_ptr.  It was included in a large number
of files as

#include <tr1/memory>

which is correct assuming that the compiler-supplied
headers include the file.  If not, the boost version can
be used, but it must be included as

#include <memory>

On ranger, when building with Intel 10.1 (the default
Intel compiler on that system), the Peridigm build fails
because tr1/memory is not found.

In this checkin, a large number of the #include <tr1/memory>
statements have been commented out completely.  It turns
out that including this include statement in three header
files is sufficient.  Things could be changed back to
explicitly include this file everywhere it is used, if
this was the desired coding style.

In the files where the include statement remains, it
has been replaced with #include "utilities/MemoryInclude.h".
This new file contains a series of preprocessor commands
to ensure that the optimal version of memory.hpp is
included, depending on what is available.  This should
significantly improve portability.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@963 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: a5839258509371997400519167194ae564a4be92