Removed commented-out code related to #including memory.hpp, which is the header for smart_ptr.
Modified the new header file that finds the best
version of memory.hpp available and includes it.
Inserted our blurb and removed the preprocessor
commands related to the Microsoft compilers.

Removed -lgfortran from the CMakeLists.txt file.
This is a compiler-specific library and should not
be in the main CMakeLists.txt file.  It caused some
issues when building with the Intel compilers (it
could not be found, no surprise there).

IMPORTANT:  The removal of -lgfortran from CMakeLists.txt
might break your build!  If this happens, you'll have to
include this library in your local cmake configuration
script.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@965 d9297710-12f9-482b-b11f-fb677dc88bbb