Added a second version of the disk impact simulation to the examples directory.
The new version is called disk_failure and
is a work in progress.  The final version of
this example will have an elastic sphere with
no damage model impacting an elastic-plastic
disk with a critical stretch failure model.

Currently Peridigm is not able to carry out
the simulation described above because not all
the pieces are there to handle the multiple
material blocks.

The input deck was written for a single material
model and no damage because there is currently no
way to specify multiple materials in the input
deck.  Plans for supporting multiple materials in
the input deck are up on the wiki.

The mesh file contains two blocks, one for the
disk and one for the sphere.  The simulation
runs correctly (matches the disk_impact example),
but only one of the blocks is output (the disk).
In its current form, disk_failure should match
exactly with disk_impact once multi-block output
functionality is available.

This example provides a means for test-driven
development in the short term, and will provide
a nice example problem in the long term.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@971 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 50eb8c2a8883891547f89fd94d9ea18f8a640d47