Added a wave propagation regression test.  The test is comprised of a bar with initial velocities applied to one end.  The wave travels the distance of the bar, reflects, and travels about halfway back.
I intend to add a second version of this test,
the only difference being that the bar will be
broken into a handful of arbitrary blocks.  The
second version of the test should yield the same
result as the current, single-block version (will
test against the exact same gold files).



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@972 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 77c75299ad8a690f84b3e9f6f81b3bb734583a72