Added multi-block version of WaveInBar test.
The test is currently deactivated in test/regression/CMakeLists.txt
pending more work on the IO system to handle multiple blocks.

The test is the same as WaveInBar, except that the bar is broken
up into eight arbitrary blocks.  The gold files for this test are
the same gold files that are used in the one-block version of the
test (with the file names changed, of course).  In this way the test
tests if there is any change in the simulation (within a tolerance)
due to the use of multiple blocks.

Currently only the first block is output.  The output for this block
appears to match the corresponding output in the single-block case,
which is encouraging.

A small change was made to Peridigm_OutputManager_VTK_XML.cpp to
avoid a std::vector access error.  The error manifested in the
four-processor version of the test, due to the fact that some blocks
had no nodes on some of the processors.  Hence a vector was being
allocated with size zero, and a subsequent call to ->at(0) failed.

Once the IO system is upgraded to handle multiple blocks, this test
should be turned on and should (hopefully) pass as is.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@974 d9297710-12f9-482b-b11f-fb677dc88bbb