Switched the linear solver for quasi-statics from AztecOO to Belos.
In the process, corrected some inconsistencies regarding the
maps used to create the tangent matrix and global vectors.
The issue is that the tangent matrix is an Epetra_FECrsMatrix,
which can only be constructed with an Epetra_Map.  The main
set of global vectors in Peridigm are built on Epetra_BlockMaps,
and are therefore not compatible with the tangent.  Changes
have been made such that the vectors which are used directly
with the tangent are now created with the same Epetra_Map that
the tangent is created with.  This has the unfortunate side
effect of having to move data between these vector and the
main Peridigm vectors manually (i.e., the Update() function
is not called).



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@993 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 0b09207876a203f7844fe6be52c463eeeb11f70c