Initial implementation of Jacobian calculation using automatic differentiation (Sacado package) for the peridynamic linear solid material model.
Two functions in the material_evaluation directory
have been duplicated, the new versions have "AD" at
the end of the name and are templated to enable the
use of automatic differentiation:

computeDilatationAD()
computeInternalForceLinearElasticAD()

There really is no need to keep the non-templated
versions around, but they have been left there for
the time being to allow for additional testing.

The actual construction of the tangent via calls
to the templated constitutive-model functions takes
place in

computeAutomaticDifferentiationJacobian()

in the LinearElasticIsotropicMaterial class.

The initial implementation uses an "element-by-element"
approach, very similar to what is done for the
finite-difference Jacobian.  Use of Sacado::Fad::DFad
on the entire model would result in massive (unacceptable)
memory allocations.  More investigation is needed to
examine other possible options.

Initial benchmarking suggests that the automatic
differentiation Jacobian is roughly twice as
fast as central-difference probing, and rougly
25% faster than forward-difference probing.  These
numbers are for only the portion of the code in which
the tangent is constructed, they are not total run
time for the analysis.

Convergence behavior is virtually identical for the
automatic differentiation Jacobian and the central-difference
Jacobian for the linear peridynamic solid material model.

Much more testing to come...



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@995 d9297710-12f9-482b-b11f-fb677dc88bbb