This is a first checkin for enabling user-defined functions to set the initial velocities and displacements of nodesets.
The only change to the input deck is in the format of the value set for each nodeset. In the old format, we would say
<ParameterList name="Initial Velocity Min X Face">
  <Parameter name="Type" type="string" value="Initial Velocity"/>
  <Parameter name="Node Set" type="string" value="Min X Node Set"/>
  <Parameter name="Coordinate" type="string" value="x"/>
  <Parameter name="Value" type="double" value="1.0"/>
</ParameterList>

... and now we say 

<ParameterList name="Initial Velocity Min X Face">
  <Parameter name="Type" type="string" value="Initial Velocity"/>
  <Parameter name="Node Set" type="string" value="Min X Node Set"/>
  <Parameter name="Coordinate" type="string" value="x"/>
  <Parameter name="Value" type="string" value="1.0"/>
</ParameterList>

(In the "Value" entry, the type of the parameter has changed from "double" to "string".)

In the "Value" entry one can now enter expressions, such as 

x + y + z (where x,y,z are taken to mean the initial position of each particle)

or

sqrt(abs(x)) + y * z

or anything similar. 

Eventually we will define a separate section of the input deck where the user can specify functions,
and the "Initial Velocity" ParameterList will refer back to those functions. 



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1001 d9297710-12f9-482b-b11f-fb677dc88bbb