Initial implementation of partial volume calculation.
Totally untested.

Also started work on a compute class to calculate the
neighborhood volume.  This will provide a check on the
partial volume calculation.

The compute class interface needs to be updated such
that the neighborhood list is passed in (so that bond
data can be accessed).  It would also be good to add
initialization and setup methods, or something along
those lines.  I'll hopefully get to this soon.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1041 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: cd9e9036b3c864a8ada0a430146b6f954cbaac6d