Temporary fix to shear correction factor, sets the correction factor to one over its current value.
A more permanent solution is forthcoming and will be included
in a refactor of the material models.  The refactor will 1) fix
and rename the dsf, 2) add partial volumes, 3) add an influence
function, 4) remove non-templated versions of functions in cases
where there is currently both a templated and a non-templated
version of the function.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1043 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 5a4ae867772f63ec3b75d7d50e94de76c4e8c0dd