Initial implementation of critical time step calculation for explicit time integration. The approximate critical time step is computed using the assumption of a linear bond-based material model.
This functionality is not yet tested.

The user-supplied time step ("Fixed dt" in the Verlet
parameter list), if present, overrides the computed
critical time step.

A safety factor may be supplied in the Verlet
parameter list.  The safety factor is denoted
in the parameter list as "Safety Factor" and is
applied to the time step, regardless of whether
it is computed or user supplied.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1064 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 0055abba345deb95861d57d939d9aafd109866c6