Added a predictor to the quasi-static time integrator.
On the first nonlinear iteration of each load step (except
the first load step), the nodal positions are predicted
based on the velocities from the previous load step.

As part of this work, velocities are now stored and output.
This required re-blessing of the gold files for the
quasi-statics tests.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1069 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: cbaa84c7fc4dc85ccaa6f155a43ff6a8aa68cd4b