Added the '-stat' option to the exodiff commands in the python scripts that run the tests.  This flag is required to get exodiff to return a nonzero error code in the event that the files do not match.  Without this flag, exodiff returns a nonzero error code only if exodiff itself fails to run properly.
Now that this flag is turned on, the WaveInBar_MultiBlock_np4
test is failing.  It appears that this test was failing
previously (exodiff was finding significant diffs), but
ctest was showing 'pass' because the actual error code
returned by exodiff was zero.

Also, added a two-block version of the simple tensile
bar test (tests multiple blocks with quasi-statics).




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1122 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: e343f559fb497e469a45a8ec75913fad847c25a4