Bug fix in ExodusII writer.  The four processor version of WaveInBar_MultiBlock now passes and has been turned back on.
It's not clear if this is a bug fix or more of a workaround.
In the WaveInBar_MultiBlock test, the blocks were not being
output in the assumed order.  This assumption has been removed
from the creation of the exodus elem_map, which now matches
correctly with the element output order.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1125 d9297710-12f9-482b-b11f-fb677dc88bbb