Incremental checkin toward multiple material functionality.
The input deck has been refactored.  An example of the new
format is in the test Bar_TwoBlocks_TwoMaterial_QS.

The code and tests build, but a large number of the tests
fail.  These tests have been temporarily disabled and will
be updated and turned back on one at a time.

Once the tests are passing it will be time to refactor
the contact section of the input deck to facilitate
the assignment of contact models to pairs of blocks.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1137 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: b01b1b6fce269b3c9d95314c3d558b4390081fa4