Updated contact section of input deck.
The new input-deck format represents the
functionality that is intended to go into Peridigm.
That is, the ability to specify multiple contact models
and to assign them to different interactions as
defined by a pair of blocks (including self contact
when a single block is duplicated to create the
pair of blocks).

Currently, Peridigm only supports one contact model
per analysis, and that contact model is assigned to
all interactions (i.e., it is on everywhere). A
significant refactor of the contact code is required
to enable the functionality described above.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1148 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: c30a9ecfad49dc11c33297d40b67ecd8c4b9c4d2