Decoupled damage models and material models.
Damage models are now specified in a Damage Models
section of the input deck.  They are associated with
specific blocks in the Blocks section.

Bond damage is now computed in a stand-alone evaluator.
References to damage models have been removed from the
material models.

The way things are currently set up, damage models are
evaluated prior to material models.  This is correct for
the critical stretch damage model (the only one currently
implemented), but it may cause issues in the future if
a yet-to-be-implemented damage model depends on material
state (which is updated by the material model).  This
won't be to hard to deal with, we'll cross that bridge
when we get there.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1152 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: f418d9d09492d44a98ee41b4522f8c4955eea8cd