Corrected the order of the arguments supplied to exodiff by the test scripts.
Older versions of exodiff were apparently
forgiving when given arguments in the incorrect
order.  Specifically, we had "-f filename.exodiff"
at the end of the argument list, and it needs to
appear prior to the names of the files that are
to be compared.  Supplying an argument after the
file names triggers the creation of a diff.e file,
where the name of the diff file is the final argument.
So, with the newest version of exodiff, we were
getting a diff file called "-f" and at least one
of the test was failing (despite Peridigm performing
properly).



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1158 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: e06c70264014d6f8fc99540cc59c1191900ca163