Code clean-up for release:
DSF has been renamed SHEAR_CORRECTION_FACTOR.

The material parameter previously called "Test"
has been renamed "Disable Plasticity" and is now
available for both the elastic-plastic model and
the elastic-plastic model with hardening.

Added option to turn on and off the shear correction
factor.  The option is called "Apply Shear Correction
Factor" and is available for the elastic-plastic model
and the elastic-plastic model with hardening.  The shear
correction factor still needs to be implemented for
the elastic and visco-elastic material models.

Added option for specifying either a finite-difference
Jacobian or an automatic-differentiation Jacobian.  The
new material-model variable is called "Apply Automatic
Difference Jacobian" and defaults to true.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1160 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: ca97d518ba519888c8a88fc9b0e541784e28c295