Removed timers from material models, all timers are now called in Peridigm.cpp.
If there is a timer in a material model,
and there are multiple blocks with different
materials, you can end up in a situation where a
given processor is only calling a subset
of the material models.  As expected, this
causes MPI errors with the timers if they
are called within the material models themselves.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1168 d9297710-12f9-482b-b11f-fb677dc88bbb