Added relative residual option to quasi-static solver.
The user may specify either "Absolute Tolerance" or
"Relative Tolerance".  The relative tolerance is
relative to the L2 norm of the reaction vector that
results from the applied increment in kinematic B.C.
at each load step.  Hence the relative tolerance
adjusts as the simulation progresses.

This is not hooked up for NOX, only for the original
Peridigm quasi-static solver.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1207 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 692bb691d32aa08ac02012ac0e40672007c53e43