Changed the way the tangent is modlfied for kinematic boundary conditions.
Originally, the rows and columns corresponding
to the kinematic B.C. were set to zero and a 1.0
was assigned to the diagonal.  Now, the value
assigned to the diagonal is equal to -1.0 times
the L2 norm of the diagonal (prior to the application
of kinematic B.C.).  The -1.0 is used because
the entries on the diagonal are all negative as a
result of the internal force sign convention.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1214 d9297710-12f9-482b-b11f-fb677dc88bbb