Updates to NOX-based quasi-static solver:
Created a function called jacobianDiagnostics()
that can be called to monitor the state of the
Jacobian as the solver iterates.  Currently,
it computes and prints to screen the condition number
at each nonlinear iteration.  It is activated
by the user setting a parameter called
"Jacobian Diagnostics" to true in the "Solver"
parameter list.  The plan is to expand this
function to include other checks, such as
symmetry.

Replace call to solver->solve() with a manual
loop over calls to solver->step().  This provides
more control over the solve and allows for printing
to the screen, etc.

Elliminated a ton of parameter list manipulation
at the top of executeNOXQuasiStatic().  The
parameter lists are now just read directly from
file.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1225 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: d9da85e7c96ec73c228de071f46109cd1cb762c5