Regression test for default block functionality.
This will allow user to specify a "default" blocktype in the input deck which will be assigned to all blocks whose type is not explicitly specified. This is useful when we have many hundreds or thousands of blocks of the same type and don't want to list all the blocks explicitly in the input deck. 




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1239 d9297710-12f9-482b-b11f-fb677dc88bbb