Changed plastic yield value calculation.
The yield value calculation took an input yield stress that was really the "shear
yield stress" which is rarely published and not what people generally associate with
yield stress.  This was changed to be consistent with the yield stress measured in a
standard quasi-static tension test.

Also, removed the relative path information from the calls to epu and conjoin in 
MergeFiles.py, this will require these to be in the User's PATH, but should be more
flexible when different versions of these programs are installed.


git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1240 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: d23ac564e6e7f7bc3478c00d3a613792ab301059