Implemented ability to pass a parameter list from the input deck to a compute class.
To do this, create a "Compute Class Parameters"
parameter list at the main level in the input
deck.  For example, the following entry in the
input deck would cause the "Test Case" parameter
list to be passed to an instance of the
Number_Of_Neighbors compute class.

<ParameterList name="Compute Class Parameters">
  <ParameterList name="Test Case">
    <Parameter name="Compute Class" type="string" value="Number_Of_Neighbors"/>
    <Parameter name="Test" type="string" value="This is a test."/>
  </ParameterList>
</ParameterList>

Note that compute classes are now instantiated
based on either an entry in the "Compute Class
Parameters" parameter list, or on an entry in
the "Output Variables" parameter list in the
Output section of the input deck.  Output to
exodus, however, is triggered based only on
entries in the Output section.

Note also that multiple instances of a compute
class may be instantiated though the "Compute
Class Parameters" parameters list.  This allows
for, say, multiple instances of the proposed
"Nearest Point Data" compute class.  It does
introduce some risk, however, because if a compute
class writes to a hard-coded output variable,
instantiating multiple instances of a class could
lead to double-counting errors.




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1264 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 16cb0c63eeec1df6fcb4d68b43d97656e724d4f8