Refactored storage of global data.
The Block and DataManager interfaces for
accessing global data are now the same
as the interfaces for accessing other
types of data.

Global data are stored as static members
of the DataManger class.  Thus, on a
given processor, accessing global data
from any Block will yield the same
Epetra_Vector.  The Epetra_Vector will
be of length one for scalar data and
length three for vector data.

Note that syncing across processors is
not done anywhere behind the scenes and
is the reponsibility of the user of the
global data.




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1270 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 4dcdce58c368b3297f55ad21b9a9e4a75be8b312