Refactored the ComputeManager constructor such that compute classes corresponding to requested Output Variables are constructed first, followed by compute classes that are requested via Compute Class Parameters.
The upshot of this is that compute classes
such as Nearest_Point_Data can operate on
derived quantities (i.e., quantities
computed by other compute classes).

The order of evaluation of compute classes
is based on the order in which they appear
in the input deck.  Compute classes corresponding
to the Output Variables section are evaluated
first (in the order they appear), followed
by the compute classes in the Compute Class
Parameters list (again, in the order they appear).

If a compute class appears in both the Output
Variables list and the Compute Class Parameters
list, the entry in the Compute Class Parameters
list takes precedence.  Currently, Peridigm
does not have a compute class for which this
issue could arise, but it may come up down
the road.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1275 d9297710-12f9-482b-b11f-fb677dc88bbb