Minor modifications to tensile_test example problem.
Added a few files for post-processing.  The
file tensile_test.post is a shell script that
runs blot to extract data from the exodus
output file (must run epu first if running
in parallel) and then runs gnuplot to generate
an engineering stress versus engineering strain
curve.

The simulation does a poor job of recovering
the elastic modulus because 1) the discretization
is very coarse, and 2) the surface correction
factor is not currently implemented for the
elastic model.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1276 d9297710-12f9-482b-b11f-fb677dc88bbb