Initial commit for influence function implementation.
Right now there is a place holder function called 
scalarInfluenceFunction() that returns 1.0.  It is inside 
material_utilities.cxx and was implemented in the elastic material 
model to ensure all the unit tests still passed.  The next check-in
will implement in the other material models, followed by implementation
of a 1/r style influence function for additional testing.  The longer
term goal is to allow the user to specify their own influence function
in the input deck.  

Another thing to consider would be to implement an "influence state"
that would have a data structure similar to bondDamage and would allow
for more complex influences and could consume the bondDamage as well.
I think this will be necassarry to get the correct stiffness (alpha)
around evolving cracks where the weighted volume (including 
the omega term) should really be computed every time step.


git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1286 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 4ce2da1a6e1bdc5609d6fec046e4f14174d6a932