Updated several varients of the Bar test.
The horizon in these tests was set to exactly
3.0 times the mesh spacing, which results in
nodes being exactly on the edge of the horizon.
The horizon for these tests is not set to 3.015
times the horizon to avoid this situation.

Also, added Number_Of_Neighbors to several tests
to provide a slightly better check on the neighbor
lists.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1290 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 4f4db9e81f758f9c2bcbcc69cf06246792003ef1