Altered build so that VTK is an optional dependency.
Woo-hoo!

Thanks John Mitchell and Karen Devine for all the help
with this.

To build with VTK, give cmake the option

-D USE_VTK:BOOL=ON

If VTK is used, Peridigm and all the tests will use
the VTK implementation of the K-D tree.  If VTK is
not used, Peridigm will use the Zoltan implementation
of the search tree.  Switching to John Mitchell's K-D tree
can be done at line 586 of NeighborhoodList.cxx.  In
the near future I'll implement the ability to switch
between the search tree implementations via the input
deck.

Some details:

Several unit tests write data to vtk files for visualization
purposes.  This output does not appear to be a key part of the
unit tests themselves.  As of this check-in, the vtk files
will be output only if the code is built with VTK.  The unit
tests are: ut_VolumeFraction, ut_ringCellVolumeFraction,
ut_QuickGrid_solidCylinder_np1, ut_QuickGrid_solidCylinder_np2,
utPdVtkPlane.

The utilities in PdVTK are available only if built against VTK.

The call to VTK's FinitePlane::bondIntersectInfinitePlane()
in BondFilter.cxx has been replaced with a home-rolled function.

The utPdVtkKdTree unit test is built and ran only if VTK
is available.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1298 d9297710-12f9-482b-b11f-fb677dc88bbb