Altered Exodus output such that databases containing only global data will be output as *.h files (h is for history).
These files do not contain the full mesh, and
instead contain a dummy mesh consisting of a
single node and single element at the origin.
Furthermore, databases containing only global
data are output only on the root processor and
are not affected by rebalance (i.e., rebalance
will not create a new .h file with the -s filename
convention as is done with .e files).

The user interface in the input deck is
unchanged.  If the user specifies an output
database that contains only global data, the
history file functionality will kick in
automatically.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1301 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: a1c98178ca16d9182d84216e32bf34de83d54b26